Change linter to Ruff
