+++ tests/test_event_trace_header.py
@@ -1,25 +1,27 @@
+from __future__ import annotations
+
 from unittest.mock import Mock
 
 from dissect.etl.headers.headers import EventTraceHeader, Marker
 
 EVENT_TRACE_EVENT = (
-    b"\x12\x01\x13\xC0\x01\x00\x00\x00\xE8\x02\x00\x00\xE4\x02\x00\x00"
-    b"\x2B\x15\x41\x0B\x00\x00\x00\x00\xA7\xCE\xDD\xB8\x20\xB5\x09\x49"
-    b"\xBC\xEB\xE0\x17\x0C\x9F\x0E\x99\x00\x00\x00\x0B\x05\x00\x00\x00"
+    b"\x12\x01\x13\xc0\x01\x00\x00\x00\xe8\x02\x00\x00\xe4\x02\x00\x00"
+    b"\x2b\x15\x41\x0b\x00\x00\x00\x00\xa7\xce\xdd\xb8\x20\xb5\x09\x49"
+    b"\xbc\xeb\xe0\x17\x0c\x9f\x0e\x99\x00\x00\x00\x0b\x05\x00\x00\x00"
     b"\x00\x00\x00\x00\x00\x20\x00\x00\x11\x00\x00\x00\x02\x00\x00\x00"
     b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-    b"\x48\x00\x0C\x00\x01\x00\x3D\x00\x3D\x00\x4D\x69\x63\x72\x6F\x73"
-    b"\x6F\x66\x74\x2E\x57\x69\x6E\x64\x6F\x77\x73\x2E\x53\x65\x72\x76"
-    b"\x69\x63\x65\x43\x6F\x6E\x74\x72\x6F\x6C\x4D\x61\x6E\x61\x67\x65"
-    b"\x72\x00\x13\x00\x01\x1A\x73\x50\x4F\xCF\x89\x82\x47\xB3\xE0\xDC"
-    b"\xE8\xC9\x04\x76\xBA\x00\x00\x00\x38\x00\x0B\x00\x00\x00\x2C\x00"
-    b"\x2C\x00\x00\x53\x65\x72\x76\x69\x63\x65\x48\x6F\x73\x74\x53\x74"
-    b"\x61\x72\x74\x65\x64\x00\x53\x65\x72\x76\x69\x63\x65\x48\x6F\x73"
-    b"\x74\x4E\x61\x6D\x65\x00\x01\x50\x49\x44\x00\x08\x00\x00\x00\x00"
-    b"\x43\x00\x3A\x00\x5C\x00\x57\x00\x69\x00\x6E\x00\x64\x00\x6F\x00"
-    b"\x77\x00\x73\x00\x5C\x00\x73\x00\x79\x00\x73\x00\x74\x00\x65\x00"
-    b"\x6D\x00\x33\x00\x32\x00\x5C\x00\x73\x00\x70\x00\x70\x00\x73\x00"
-    b"\x76\x00\x63\x00\x2E\x00\x65\x00\x78\x00\x65\x00\x00\x00\xF4\x01"
+    b"\x48\x00\x0c\x00\x01\x00\x3d\x00\x3d\x00\x4d\x69\x63\x72\x6f\x73"
+    b"\x6f\x66\x74\x2e\x57\x69\x6e\x64\x6f\x77\x73\x2e\x53\x65\x72\x76"
+    b"\x69\x63\x65\x43\x6f\x6e\x74\x72\x6f\x6c\x4d\x61\x6e\x61\x67\x65"
+    b"\x72\x00\x13\x00\x01\x1a\x73\x50\x4f\xcf\x89\x82\x47\xb3\xe0\xdc"
+    b"\xe8\xc9\x04\x76\xba\x00\x00\x00\x38\x00\x0b\x00\x00\x00\x2c\x00"
+    b"\x2c\x00\x00\x53\x65\x72\x76\x69\x63\x65\x48\x6f\x73\x74\x53\x74"
+    b"\x61\x72\x74\x65\x64\x00\x53\x65\x72\x76\x69\x63\x65\x48\x6f\x73"
+    b"\x74\x4e\x61\x6d\x65\x00\x01\x50\x49\x44\x00\x08\x00\x00\x00\x00"
+    b"\x43\x00\x3a\x00\x5c\x00\x57\x00\x69\x00\x6e\x00\x64\x00\x6f\x00"
+    b"\x77\x00\x73\x00\x5c\x00\x73\x00\x79\x00\x73\x00\x74\x00\x65\x00"
+    b"\x6d\x00\x33\x00\x32\x00\x5c\x00\x73\x00\x70\x00\x70\x00\x73\x00"
+    b"\x76\x00\x63\x00\x2e\x00\x65\x00\x78\x00\x65\x00\x00\x00\xf4\x01"
     b"\x00\x00\x00\x00\x00\x00\x00\x00"
 )
 
@@ -38,6 +40,6 @@ def test_event_header() -> None:
     assert header.ThreadId == 0x2E8
     assert header.ProcessId == 0x2E4
     assert header.TimeDelta == 0x0B41152B
-    assert header.ProviderId == b"\xA7\xCE\xDD\xB8\x20\xB5\x09\x49\xBC\xEB\xE0\x17\x0C\x9F\x0E\x99"
+    assert header.ProviderId == b"\xa7\xce\xdd\xb8\x20\xb5\x09\x49\xbc\xeb\xe0\x17\x0c\x9f\x0e\x99"
     assert header.KernelTime == 0xB000000
     assert header.UserTime == 0x5
