+++ tests/test_system_header.py
@@ -1,14 +1,19 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
 from unittest.mock import Mock
 
 import pytest
 
 from dissect.etl.exceptions import InvalidMarkerError
-from dissect.etl.headers.headers import Header
 from dissect.etl.headers.utils import select_event_header
 
+if TYPE_CHECKING:
+    from dissect.etl.headers.headers import Header
+
 SYSTEM_EVENT = (
-    b"\x02\x00\x02\xC0\x44\x00\x05\x00\xB4\x00\x00\x00\x04\x00\x00\x00"
-    b"\x8F\x8C\x29\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00"
+    b"\x02\x00\x02\xc0\x44\x00\x05\x00\xb4\x00\x00\x00\x04\x00\x00\x00"
+    b"\x8f\x8c\x29\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\x00\x00\x00"
     b"\x07\x03\x00\x00\x04\x01\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
     b"\x00\x00\x02\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
     b"\x50\x00\x00\x00\x00\x00\x00\x00"
