+++ tests/test_wmi_buffer_header.py
@@ -1,14 +1,16 @@
+from __future__ import annotations
+
 from unittest.mock import Mock
 
 from dissect.cstruct import cstruct
 
-from dissect.etl.utils import etl_def
+from dissect.etl.c_etl import etl_def
 
 RAW_BUFFER_HEADER = (
-    b"\x00\x00\x10\x00\xA8\x02\x00\x00\xA8\x02\x00\x00\x00\x00\x00\x00"
+    b"\x00\x00\x10\x00\xa8\x02\x00\x00\xa8\x02\x00\x00\x00\x00\x00\x00"
     b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
     b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x12\x00\x03\x00\x00\x00"
-    b"\xA8\x02\x00\x00\x21\x00\x04\x00\x00\x00\x00\x00\x00\x00\x00\x00"
+    b"\xa8\x02\x00\x00\x21\x00\x04\x00\x00\x00\x00\x00\x00\x00\x00\x00"
     b"\x00\x00\x00\x00\x00\x00\x00\x00"
 )
